+++ src/main/java/com/strangegrotto/wealthdraft/assets/definition/CustomTagDefinition.java
@@ -4,14 +4,13 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.strangegrotto.wealthdraft.WealthdraftImmutableStyle;
 import org.immutables.value.Value;
 
+import java.util.Optional;
 import java.util.Set;
 
 @WealthdraftImmutableStyle
 @Value.Immutable
 @JsonDeserialize(as = ImmCustomTagDefinition.class)
 public interface CustomTagDefinition {
-    // TODO Implement default values
-
     @Value.Default
     default boolean isRequired() {
         return false;
@@ -19,4 +18,6 @@ public interface CustomTagDefinition {
 
     // An empty set means all values are allowed
     Set<String> getAllowedValues();
+
+    Optional<String> getDefaultValue();
 }